refactored test cfgs
