+++ tests/macos.rs
@@ -1,3 +1,4 @@
+#![cfg(target_os = "macos")]
 use dylink::*;
 
 #[test]
