+++ tests/unix.rs
@@ -1,3 +1,5 @@
+#![cfg(unix)]
+
 #[cfg(not(target_os = "aix"))]
 #[test]
 fn test_unix_sym_info() {
