+++ tests/windows.rs
@@ -1,3 +1,4 @@
+#![cfg(windows)]
 use dylink::*;
 
 static KERNEL32: sync::LibLock = sync::LibLock::new(&["Kernel32.dll"]);
